feat: edit smart shelf rules after creation (#987)
Smart shelves previously only displayed their rules read-only on the
detail page; changing a smart query meant delete + recreate. This adds an
"Edit rules" action that opens a prefilled rule builder and saves via the
existing update_shelf path.

- Extract the create modal's condition-row editor + live-preview pane into
  a shared `shelf_rule_builder` module (`RuleDraft` + `RuleBuilder`), so
  create and edit share one implementation. `RuleDraft::from_rule` reverses
  the value encoding to prefill an editor from a saved rule.
- Add `EditShelfRulesModal`, mounted from both the web `ShelfHeader` actions
  menu and the mobile shelf-detail actions menu (smart shelves only). Saving
  calls `data::update_shelf` with `UpdateShelfRequest { rules, match_mode }`
  and bumps the parent reload signal so membership refetches immediately.
- Add a db-layer test confirming an `update_shelf` rules/match_mode change
  recomputes membership for an existing smart shelf.
- Add Playwright coverage editing a smart shelf's rule and observing the
  member grid change.
