+++ frontend/src/components/create_shelf_modal.rs
@@ -7,96 +7,13 @@
 
 use dioxus::prelude::*;
 use omnibus_shared::{
-    CreateShelfRequest, EbookMetadata, MatchMode, RuleField, RuleOp, RulePreview, Shelf, ShelfKind,
-    ShelfRule, Visibility,
+    CreateShelfRequest, EbookMetadata, MatchMode, Shelf, ShelfKind, ShelfRule, Visibility,
 };
 
 use crate::components::cover_tile::toggle_picked;
+use crate::components::shelf_rule_builder::{RuleBuilder, RuleDraft};
 use crate::components::{CoverTile, CoverTileKind};
 use crate::{data, use_server_url};
-
-/// Field/op/date-unit choices exposed in the smart-rule editor. `Status` is
-/// deliberately omitted — unsupported in v1.
-const FIELDS: &[(RuleField, &str)] = &[
-    (RuleField::Tag, "Tag"),
-    (RuleField::Author, "Author"),
-    (RuleField::Series, "Series"),
-    (RuleField::Rating, "Rating"),
-    (RuleField::Format, "Format"),
-    (RuleField::Year, "Year"),
-    (RuleField::DateAdded, "Date added"),
-    (RuleField::DateUpdated, "Date updated"),
-];
-
-const OPS: &[(RuleOp, &str)] = &[
-    (RuleOp::Is, "is"),
-    (RuleOp::IsNot, "is not"),
-    (RuleOp::Contains, "contains"),
-    (RuleOp::StartsWith, "starts with"),
-    (RuleOp::Gte, "is at least"),
-    (RuleOp::Includes, "includes"),
-    (RuleOp::InLast, "in the last"),
-    (RuleOp::Between, "between"),
-    (RuleOp::Before, "before"),
-    (RuleOp::After, "after"),
-];
-
-/// A rule row's working state. `value`/`value2`/`unit` are the raw input
-/// strings; [`encode_value`] folds them into the wire `ShelfRule.value`.
-#[derive(Clone, PartialEq)]
-struct RuleDraft {
-    field: RuleField,
-    op: RuleOp,
-    value: String,
-    /// Second date for `Between`.
-    value2: String,
-    /// Relative-window unit (`d`/`w`/`m`/`y`) for `InLast`.
-    unit: String,
-}
-
-impl Default for RuleDraft {
-    fn default() -> Self {
-        RuleDraft {
-            field: RuleField::Tag,
-            op: RuleOp::Is,
-            value: String::new(),
-            value2: String::new(),
-            unit: "d".into(),
-        }
-    }
-}
-
-impl RuleDraft {
-    /// Fold the raw inputs into the wire `value` string per the field's op.
-    fn encode_value(&self) -> String {
-        match self.op {
-            RuleOp::InLast => format!("{}{}", self.value.trim(), self.unit),
-            RuleOp::Between => format!("{}..{}", self.value.trim(), self.value2.trim()),
-            _ => self.value.trim().to_string(),
-        }
-    }
-
-    /// `true` once every input the current op needs is filled in.
-    fn is_complete(&self) -> bool {
-        match self.op {
-            RuleOp::InLast => !self.value.trim().is_empty(),
-            RuleOp::Between => !self.value.trim().is_empty() && !self.value2.trim().is_empty(),
-            _ => !self.value.trim().is_empty(),
-        }
-    }
-
-    /// Convert to a wire rule when complete.
-    fn to_rule(&self) -> Option<ShelfRule> {
-        if !self.is_complete() {
-            return None;
-        }
-        Some(ShelfRule {
-            field: self.field,
-            op: self.op,
-            value: self.encode_value(),
-        })
-    }
-}
 
 /// Create-shelf modal. Emits the created shelf via `on_created`.
 #[component]
@@ -205,7 +122,7 @@ pub fn CreateShelfModal(on_close: EventHandler<()>, on_created: EventHandler<She
                 div { class: "shelf-modal-body",
                     match kind() {
                         ShelfKind::Smart => rsx! {
-                            SmartBody {
+                            RuleBuilder {
                                 match_mode: match_mode(),
                                 rules,
                                 on_match_mode: move |m| match_mode.set(m),
@@ -269,277 +186,6 @@ fn VisibilityToggle(visibility: Visibility, on_change: EventHandler<Visibility>)
                 onclick: move |_| on_change.call(Visibility::Public),
                 "Public"
             }
-        }
-    }
-}
-
-/// Smart-rule editor + live-preview pane.
-#[component]
-fn SmartBody(
-    match_mode: MatchMode,
-    rules: Signal<Vec<RuleDraft>>,
-    on_match_mode: EventHandler<MatchMode>,
-    server_url: String,
-) -> Element {
-    let mut rules = rules;
-    let preview = use_signal(|| None::<RulePreview>);
-
-    // Recompute the preview whenever the rule set or match mode changes. Keyed
-    // on a memo of the encoded rules so unrelated re-renders don't refetch.
-    let preview_key = use_memo(move || {
-        let complete: Vec<ShelfRule> = rules.read().iter().filter_map(RuleDraft::to_rule).collect();
-        (match_mode, complete)
-    });
-    let preview_url = server_url.clone();
-    use_effect(move || {
-        let (mode, wire) = preview_key();
-        let url = preview_url.clone();
-        let mut preview = preview;
-        if wire.is_empty() {
-            preview.set(None);
-            return;
-        }
-        spawn(async move {
-            if let Ok(p) = data::preview_shelf_rule(&url, mode, wire).await {
-                preview.set(Some(p));
-            }
-        });
-    });
-
-    rsx! {
-        div { class: "shelf-smart",
-            div { class: "shelf-smart-editor",
-                div { class: "shelf-match-row",
-                    span { "Match" }
-                    select {
-                        class: "shelf-select",
-                        "data-testid": "shelf-match-mode",
-                        value: match_mode.as_str(),
-                        onchange: move |e| {
-                            if let Some(m) = MatchMode::from_str(&e.value()) {
-                                on_match_mode.call(m);
-                            }
-                        },
-                        option { value: "any", "any" }
-                        option { value: "all", "all" }
-                    }
-                    span { "of these conditions" }
-                }
-
-                for (i, draft) in rules.read().iter().cloned().enumerate() {
-                    ConditionRow {
-                        key: "{i}",
-                        index: i,
-                        draft: draft,
-                        can_remove: rules.read().len() > 1,
-                        on_change: move |d: RuleDraft| {
-                            rules.write()[i] = d;
-                        },
-                        on_remove: move |_| {
-                            rules.write().remove(i);
-                        },
-                    }
-                }
-
-                button {
-                    r#type: "button",
-                    class: "shelf-add-condition",
-                    "data-testid": "add-condition",
-                    onclick: move |_| rules.write().push(RuleDraft::default()),
-                    "\u{FF0B} Add condition"
-                }
-            }
-
-            div { class: "shelf-preview",
-                {render_preview(&preview.read(), &server_url)}
-            }
-        }
-    }
-}
-
-/// The live-preview pane content: count + sample covers.
-fn render_preview(preview: &Option<RulePreview>, server_url: &str) -> Element {
-    match preview {
-        None => rsx! {
-            p { class: "shelf-preview-empty mono", "Add a condition to preview matches." }
-        },
-        Some(p) => rsx! {
-            p {
-                class: "shelf-preview-count",
-                "data-testid": "rule-preview-count",
-                em { "{p.matched}" }
-                " of {p.total} match"
-            }
-            div { class: "shelf-preview-grid",
-                for book in p.sample.iter().cloned() {
-                    div {
-                        key: "{book.id}",
-                        CoverTile {
-                            book,
-                            server_url: server_url.to_string(),
-                            sizes: "120px".to_string(),
-                            kind: CoverTileKind::ReadOnly,
-                        }
-                    }
-                }
-            }
-        },
-    }
-}
-
-/// One editable smart-rule condition row.
-#[component]
-fn ConditionRow(
-    index: usize,
-    draft: RuleDraft,
-    can_remove: bool,
-    on_change: EventHandler<RuleDraft>,
-    on_remove: EventHandler<()>,
-) -> Element {
-    let field = draft.field;
-    let op = draft.op;
-
-    // On a field change, snap the op to the first one the new field accepts.
-    let d_field = draft.clone();
-    let on_field = move |e: FormEvent| {
-        if let Some(f) = RuleField::from_str(&e.value()) {
-            let mut d = d_field.clone();
-            d.field = f;
-            if !f.accepts(d.op) {
-                d.op = OPS
-                    .iter()
-                    .map(|(o, _)| *o)
-                    .find(|o| f.accepts(*o))
-                    .unwrap_or(RuleOp::Is);
-            }
-            on_change.call(d);
-        }
-    };
-    let d_op = draft.clone();
-    let on_op = move |e: FormEvent| {
-        if let Some(o) = RuleOp::from_str(&e.value()) {
-            let mut d = d_op.clone();
-            d.op = o;
-            on_change.call(d);
-        }
-    };
-    let d_val = draft.clone();
-    let on_val = move |e: FormEvent| {
-        let mut d = d_val.clone();
-        d.value = e.value();
-        on_change.call(d);
-    };
-    let d_val2 = draft.clone();
-    let on_val2 = move |e: FormEvent| {
-        let mut d = d_val2.clone();
-        d.value2 = e.value();
-        on_change.call(d);
-    };
-    let d_unit = draft.clone();
-    let on_unit = move |e: FormEvent| {
-        let mut d = d_unit.clone();
-        d.unit = e.value();
-        on_change.call(d);
-    };
-
-    rsx! {
-        div { class: "shelf-condition", "data-testid": "condition-row-{index}",
-            select {
-                class: "shelf-select",
-                "data-testid": "condition-field-{index}",
-                value: field.as_str(),
-                onchange: on_field,
-                for (f, label) in FIELDS.iter() {
-                    option { key: "{f.as_str()}", value: f.as_str(), "{label}" }
-                }
-            }
-            select {
-                class: "shelf-select",
-                "data-testid": "condition-op-{index}",
-                value: op.as_str(),
-                onchange: on_op,
-                for (o, label) in OPS.iter().filter(|(o, _)| field.accepts(*o)) {
-                    option { key: "{o.as_str()}", value: o.as_str(), "{label}" }
-                }
-            }
-            {condition_value_input(&draft, on_val, on_val2, on_unit)}
-            if can_remove {
-                button {
-                    r#type: "button",
-                    class: "shelf-condition-remove",
-                    "data-testid": "condition-remove-{index}",
-                    "aria-label": "Remove condition",
-                    onclick: move |_| on_remove.call(()),
-                    "\u{00D7}"
-                }
-            }
-        }
-    }
-}
-
-/// Field-aware value input(s) for a condition row.
-fn condition_value_input(
-    draft: &RuleDraft,
-    on_val: impl FnMut(FormEvent) + 'static,
-    on_val2: impl FnMut(FormEvent) + 'static,
-    on_unit: impl FnMut(FormEvent) + 'static,
-) -> Element {
-    let value = draft.value.clone();
-    let value2 = draft.value2.clone();
-    let unit = draft.unit.clone();
-
-    // Date fields carry op-specific inputs.
-    if draft.field.is_date() {
-        return match draft.op {
-            RuleOp::InLast => rsx! {
-                div { class: "shelf-value-group",
-                    input {
-                        r#type: "number", min: "1", class: "shelf-value shelf-value--num",
-                        placeholder: "30", value: "{value}", oninput: on_val,
-                    }
-                    select { class: "shelf-select", value: "{unit}", onchange: on_unit,
-                        option { value: "d", "days" }
-                        option { value: "w", "weeks" }
-                        option { value: "m", "months" }
-                        option { value: "y", "years" }
-                    }
-                }
-            },
-            RuleOp::Between => rsx! {
-                div { class: "shelf-value-group",
-                    input { r#type: "date", class: "shelf-value", value: "{value}", oninput: on_val }
-                    span { class: "shelf-value-sep", "to" }
-                    input { r#type: "date", class: "shelf-value", value: "{value2}", oninput: on_val2 }
-                }
-            },
-            _ => rsx! {
-                input { r#type: "date", class: "shelf-value", value: "{value}", oninput: on_val }
-            },
-        };
-    }
-
-    // Rating / Year are numeric; the rest are free text matched by name
-    // (case-insensitive), so author/series take the name the user sees.
-    let (input_type, placeholder) = match draft.field {
-        RuleField::Rating => ("number", "4"),
-        RuleField::Year => ("number", "2024"),
-        RuleField::Author => ("text", "author name"),
-        RuleField::Series => ("text", "series name"),
-        _ => ("text", "value"),
-    };
-    let step = if matches!(draft.field, RuleField::Rating) {
-        "0.5"
-    } else {
-        "1"
-    };
-    rsx! {
-        input {
-            r#type: "{input_type}",
-            step: "{step}",
-            class: "shelf-value",
-            placeholder: "{placeholder}",
-            value: "{value}",
-            oninput: on_val,
         }
     }
 }
